requirements.txt
streamlit
sympy
matplotlib
numpy
ipython
openai
python-dotenv
pillow